+++ ownCloud/Client/Actions/Actions+Extensions/UploadBaseAction.swift
@@ -18,6 +18,7 @@
 
 import UIKit
 import ownCloudSDK
+import ownCloudApp
 
 class UploadBaseAction: Action {
 
@@ -41,26 +42,30 @@ class UploadBaseAction: Action {
 
 	// MARK: - Upload
 	func upload(itemURL: URL, to rootItem: OCItem, name: String, completionHandler: UploadCompletionHandler? = nil, placeholderHandler:UploadPlaceholderCompletionHandler? = nil, importByCopy:Bool = false) {
-		if let progress = core?.importFileNamed(name,
-												at: rootItem,
-												from: itemURL,
-												isSecurityScoped: false,
-												options: importByCopy ? [OCCoreOption.importByCopying : true] : nil,
-												placeholderCompletionHandler: { (error, item) in
-													if error != nil {
-														Log.debug("Error uploading \(Log.mask(name)) to \(Log.mask(rootItem.path)), error: \(error?.localizedDescription ?? "" )")
-													}
-													placeholderHandler?(item, error)
-		},
-												resultHandler: { (error, _ core, _ item, _) in
-													if error != nil {
-														Log.debug("Error uploading \(Log.mask(name)) to \(Log.mask(rootItem.path)), error: \(error?.localizedDescription ?? "" )")
-														completionHandler?(false, item)
-													} else {
-														Log.debug("Success uploading \(Log.mask(name)) to \(Log.mask(rootItem.path))")
-														completionHandler?(true, item)
-													}
-		}) {
+		if let progress = core?.importItemNamed(name,
+							at: rootItem,
+							from: itemURL,
+							isSecurityScoped: false,
+							options: [
+								OCCoreOption.importByCopying : importByCopy,
+								OCCoreOption.automaticConflictResolutionNameStyle : OCCoreDuplicateNameStyle.bracketed.rawValue
+							],
+							placeholderCompletionHandler: { (error, item) in
+								if error != nil {
+									Log.debug("Error uploading \(Log.mask(name)) to \(Log.mask(rootItem.path)), error: \(error?.localizedDescription ?? "" )")
+								}
+								placeholderHandler?(item, error)
+							},
+							resultHandler: { (error, _ core, _ item, _) in
+								if error != nil {
+									Log.debug("Error uploading \(Log.mask(name)) to \(Log.mask(rootItem.path)), error: \(error?.localizedDescription ?? "" )")
+									completionHandler?(false, item)
+								} else {
+									Log.debug("Success uploading \(Log.mask(name)) to \(Log.mask(rootItem.path))")
+									completionHandler?(true, item)
+								}
+							}
+		) {
 			self.publish(progress: progress)
 		} else {
 			Log.debug("Error setting up upload of \(Log.mask(name)) to \(Log.mask(rootItem.path))")