+++ ownCloud/Client/Viewer/DisplayViewController.swift
@@ -49,6 +49,8 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 
 	// MARK: - Configuration
 	var item: OCItem?
+	var itemIndex: Int?
+
 	private var coreConnectionStatusObservation : NSKeyValueObservation?
 	weak var core: OCCore? {
 		willSet {
@@ -72,15 +74,26 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		}
 	}
 
+	// This shall be set to false if DisplayViewController sublass is able to handle streamed data (e.g. audio, video)
+	var requiresLocalItemCopy: Bool = true
+
 	var source: URL? {
 		didSet {
 			OnMainThread(inline: true) {
 				self.iconImageView?.isHidden = true
 				self.hideItemMetadataUIElements()
-				self.renderSpecificView()
+				self.renderSpecificView(completion: { (success) in
+					if !success {
+						self.iconImageView?.isHidden = false
+						self.infoLabel?.text = "File couldn't be opened".localized
+						self.infoLabel?.isHidden = false
+					}
+				})
 			}
 		}
 	}
+
+	var httpAuthHeaders: [String : String]?
 
 	var shallDisplayMoreButtonInToolbar = true
 
@@ -111,7 +124,7 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 	private var cancelButton : ThemeButton?
 	private var metadataInfoLabel: UILabel?
 	private var showPreviewButton: ThemeButton?
-	private var noNetworkLabel : UILabel?
+	private var infoLabel : UILabel?
 
 	// MARK: - Delegate
 	weak var editingDelegate: DisplayViewEditingDelegate?
@@ -141,10 +154,10 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		metadataInfoLabel = UILabel()
 		cancelButton = ThemeButton(type: .custom)
 		showPreviewButton = ThemeButton(type: .custom)
-		noNetworkLabel = UILabel()
+		infoLabel = UILabel()
 		progressView = UIProgressView(progressViewStyle: .bar)
 
-		guard let iconImageView = iconImageView, let metadataInfoLabel = metadataInfoLabel, let progressView = progressView, let cancelButton = cancelButton, let showPreviewButton = showPreviewButton, let noNetworkLabel = noNetworkLabel else {
+		guard let iconImageView = iconImageView, let metadataInfoLabel = metadataInfoLabel, let progressView = progressView, let cancelButton = cancelButton, let showPreviewButton = showPreviewButton, let noNetworkLabel = infoLabel else {
 			return
 		}
 
@@ -252,14 +265,23 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		self.render()
 	}
 
+	override func viewWillAppear(_ animated: Bool) {
+		super.viewWillAppear(animated)
+		updateNavigationBarItems()
+	}
+
 	func updateNavigationBarItems() {
-		if let parent = parent, let itemName = item?.name {
+		if let parent = parent, let itemName = item?.name, let queryState = query?.state {
 			parent.navigationItem.title = itemName
 
 			if shallDisplayMoreButtonInToolbar {
-				let actionsBarButtonItem = UIBarButtonItem(image: UIImage(named: "more-dots"), style: .plain, target: self, action: #selector(optionsBarButtonPressed))
-				actionsBarButtonItem.accessibilityLabel = itemName + " " + "Actions".localized
-				parent.navigationItem.rightBarButtonItem = actionsBarButtonItem
+				if queryState != .targetRemoved {
+					let actionsBarButtonItem = UIBarButtonItem(image: UIImage(named: "more-dots"), style: .plain, target: self, action: #selector(optionsBarButtonPressed))
+					actionsBarButtonItem.accessibilityLabel = itemName + " " + "Actions".localized
+					parent.navigationItem.rightBarButtonItem = actionsBarButtonItem
+				} else {
+					parent.navigationItem.rightBarButtonItem = nil
+				}
 			}
 		}
 	}
@@ -297,7 +319,7 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		}
 	}
 
-	func renderSpecificView() {
+	func renderSpecificView(completion: @escaping  (_ success:Bool) -> Void) {
 		// This function is intended to be overwritten by the subclases to implement a custom view based on the source property.s
 	}
 
@@ -307,7 +329,7 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		cancelButton?.isHidden = true
 		metadataInfoLabel?.isHidden = true
 		showPreviewButton?.isHidden = true
-		noNetworkLabel?.isHidden = true
+		infoLabel?.isHidden = true
 	}
 
 	private func render() {
@@ -318,7 +340,7 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		case .noNetworkConnection:
 			self.progressView?.isHidden = true
 			self.cancelButton?.isHidden = true
-			self.noNetworkLabel?.isHidden = false
+			self.infoLabel?.isHidden = false
 			self.showPreviewButton?.isHidden = true
 
 		case .errorDownloading, .canceledDownload:
@@ -329,14 +351,15 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		case .downloading(_):
 			self.progressView?.isHidden = false
 			self.cancelButton?.isHidden = false
-			self.noNetworkLabel?.isHidden = true
+			self.infoLabel?.isHidden = true
 			self.showPreviewButton?.isHidden = true
 
 		case .notSupportedMimeType:
 			self.progressView?.isHidden = true
 			self.cancelButton?.isHidden = true
-			self.noNetworkLabel?.isHidden = true
+			self.infoLabel?.isHidden = true
 			self.showPreviewButton?.isHidden = true
+
 		}
 	}
 
@@ -344,7 +367,7 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		self.progressView?.progress = 0.0
 		self.progressView?.isHidden = true
 		self.cancelButton?.isHidden = true
-		self.noNetworkLabel?.isHidden = true
+		self.infoLabel?.isHidden = true
 		self.showPreviewButton?.isHidden = false
 	}
 
@@ -356,11 +379,7 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 		let actionsLocation = OCExtensionLocation(ofType: .action, identifier: .moreItem)
 		let actionContext = ActionContext(viewController: self, core: core, items: [item], location: actionsLocation)
 
-		if let moreViewController = Action.cardViewController(for: item, with: actionContext, completionHandler: { [weak self] (action, _) in
-			if !(action is OpenInAction) {
-				self?.navigationController?.popViewController(animated: true)
-			}
-		}) {
+		if let moreViewController = Action.cardViewController(for: item, with: actionContext, completionHandler: nil) {
 			self.present(asCard: moreViewController, animated: true)
 		}
 	}
@@ -396,18 +415,24 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 	}
 
 	func queryHasChangesAvailable(_ query: OCQuery) {
-		query.requestChangeSet(withFlags: OCQueryChangeSetRequestFlag(rawValue: 0)) { (query, changeSet) in
+		query.requestChangeSet(withFlags: .onlyResults) { [weak self] (query, changeSet) in
 			OnMainThread {
+				Log.log("Presenting item (DisplayViewController.queryHasChangesAvailable): \(changeSet?.queryResult.description ?? "nil") - state: \(String(describing: query.state.rawValue))")
+
 				switch query.state {
 					case .idle, .contentsFromCache, .waitingForServerReply:
 						if let firstItem = changeSet?.queryResult.first {
-							if (firstItem.itemVersionIdentifier != self.item?.itemVersionIdentifier) || (firstItem.name != self.item?.name) {
-								self.present(item: firstItem)
+							if (firstItem.syncActivity != .updating) && ((firstItem.itemVersionIdentifier != self?.item?.itemVersionIdentifier) || (firstItem.name != self?.item?.name)) {
+								self?.present(item: firstItem)
+							} else {
+								self?.item = firstItem
 							}
 						}
 
-					case .targetRemoved: break
+						self?.updateNavigationBarItems()
 
+					case .targetRemoved:
+						self?.updateNavigationBarItems()
 					default: break
 				}
 			}
@@ -415,30 +440,42 @@ class DisplayViewController: UIViewController, OCQueryDelegate {
 	}
 
 	func present(item: OCItem) {
-		guard self.view != nil else {
+		guard self.view != nil, item.removed == false else {
 			return
 		}
-
-		self.item = item
-
-		metadataInfoLabel?.text = item.sizeLocalized + " - " + item.lastModifiedLocalized
 
 		switch state {
 			case .notSupportedMimeType: break
 
 			default:
-				self.stopQuery()
+				self.item = item
+				metadataInfoLabel?.text = item.sizeLocalized + " - " + item.lastModifiedLocalized
+
+				Log.log("Presenting item (DisplayViewController.present): \(item.description)")
+
 				self.startQuery()
 
-				if core?.localCopy(of: item) == nil {
-					self.downloadItem(sender: nil)
-				} else {
-					if let core = core, let file = item.file(with: core) {
-						self.source = file.url
+				if source == nil {
+					if requiresLocalItemCopy {
+						if core?.localCopy(of: item) == nil {
+							self.downloadItem(sender: nil)
+						} else {
+							if let core = core, let file = item.file(with: core) {
+								self.source = file.url
+							}
+						}
+					} else {
+						core?.provideDirectURL(for: item, allowFileURL: true, completionHandler: { (error, url, authHeaders) in
+							if error == nil {
+								self.httpAuthHeaders = authHeaders
+								self.source = url
+							}
+						})
 					}
 				}
 
-				updateNavigationBarItems()
+				self.updateNavigationBarItems()
+
 		}
 	}
 }
@@ -459,7 +496,7 @@ extension DisplayViewController : Themeable {
 		cancelButton?.applyThemeCollection(collection)
 		metadataInfoLabel?.applyThemeCollection(collection)
 		showPreviewButton?.applyThemeCollection(collection)
-		noNetworkLabel?.applyThemeCollection(collection)
+		infoLabel?.applyThemeCollection(collection)
 		self.view.backgroundColor = collection.tableBackgroundColor
 	}
 }