+++ ownCloud/Tools/VendorServices.swift
@@ -138,7 +138,7 @@ extension VendorServices : OCClassSettingsSupport {
 
 	static func defaultSettings(forIdentifier identifier: OCClassSettingsIdentifier) -> [OCClassSettingsKey : Any]? {
 		if identifier == .app {
-			return [ .isBetaBuild : false, .showBetaWarning : false, .enableUIAnimations: true ]
+			return [ .isBetaBuild : true, .showBetaWarning : true, .enableUIAnimations: true ]
 		}
 
 		return nil